backend/requirements.txt
# FastAPI y servidor
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Base de datos
sqlalchemy==2.0.36
psycopg[binary]==3.2.13
alembic==1.14.0

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.1

# Procesamiento de documentos
python-docx==1.1.2
PyPDF2==3.0.1
python-pptx==1.0.2
openpyxl==3.1.5
pandas==2.2.3

# IA y NLP
google-generativeai==0.8.3

# Google Drive
google-api-python-client==2.154.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Utilidades
pydantic[email]==2.10.3
pydantic-settings==2.6.1
psutil==6.1.0
